requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.105.0
h11==0.14.0
idna==3.6
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic-core==2.14.5
python-multipart==0.0.6
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
typing-extensions==4.9.0
urllib3==2.1.0
uvicorn==0.24.0.post1
